requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.49
PyYAML==6.0.1
watchdog==3.0.0
